Add PDF generating capability for send functionality
Using the PDFMake library to generate a pdf from the selected
transactions. The original plan was to attach all the transaction
details into the body of an email triggered with a mailto link, but the
spec for mailto: only allows plaintext in the body (meaning the receipt
images appear as giant dataURI strings; no way to set them as
attachments) and there's a limit to the character length allowed in
URL's that we would certainly run into. The switch to PDF generation
allows the pdf to be sent from within the browser (using mobile
browser's native behavior) and allows for the storage of the PDFs
generated as records.