Update database creation options
Add auto-compaction so that deleted documents will be "removed" (the
image attachments are large and storing them indefinitely after they've
been submitted is unnecessary). Need to make sure that auto-compaction
does in fact free the space. Also had to add a size option to request a
large memory buffer for the site (2 pictures fill up the 5mb initial
allocation) because a bug in mobile Safari prevents PouchDB from
allocating more space, regardless of the user's answer to the space
prompt. This issue will likely need to be revisited in the future.